requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14.0,<3.15.0
gunicorn>=20.1.0,<21.0.0
psycopg2>=2.9.0,<3.0.0